Scaffold scheduler and node-daemon for implementing migration
Console client asks the scheduler to migrate a container from one node
to another. The scheduler first connects to the receiving node-daemon,
asking to prepare for the migration. Then, the scheduler connects to
the sending node-daemon, asking to prepare for migration.

Everybody logs all the events, but nothing happens yet.